+++ src/components/taxonomy/taxonomy-move-copy.tsx
@@ -26,12 +26,12 @@ import {
 } from '@/frontend-node-types'
 import { getTranslatedType } from '@/helper/get-translated-type'
 import { getIconByTypename } from '@/helper/icon-by-entity-type'
+import { replacePlaceholders } from '@/helper/replace-placeholders'
+import { showToastNotice } from '@/helper/show-toast-notice'
 import {
   useCreateEntityLinkMutation,
   useDeleteEntityLinkMutation,
-} from '@/helper/mutations/taxonomyTerm'
-import { replacePlaceholders } from '@/helper/replace-placeholders'
-import { showToastNotice } from '@/helper/show-toast-notice'
+} from '@/mutations/taxonomyTerm'
 
 interface TaxonomyMoveCopyProps {
   taxonomyData: TaxonomyData
