+++ src/mutations/use-mutation-fetch.ts
@@ -1,10 +1,10 @@
 import { GraphQLError } from 'graphql'
 import { ClientError, GraphQLClient } from 'graphql-request'
 
-import { csrReload } from '../csr-reload'
-import { hasOwnPropertyTs } from '../has-own-property-ts'
-import { showToastNotice } from '../show-toast-notice'
-import { triggerSentry } from '../trigger-sentry'
+import { csrReload } from '../helper/csr-reload'
+import { hasOwnPropertyTs } from '../helper/has-own-property-ts'
+import { showToastNotice } from '../helper/show-toast-notice'
+import { triggerSentry } from '../helper/trigger-sentry'
 import { endpoint } from '@/api/endpoint'
 import { useAuthentication } from '@/auth/use-authentication'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
