+++ src/mutations/use-revision-decide-mutation.ts
@@ -2,7 +2,7 @@ import { gql } from 'graphql-request'
 import { useRouter } from 'next/router'
 import NProgress from 'nprogress'
 
-import { showToastNotice } from '../show-toast-notice'
+import { showToastNotice } from '../helper/show-toast-notice'
 import { useMutationFetch } from './use-mutation-fetch'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 import { RejectRevisionInput } from '@/fetcher/graphql-types/operations'
