+++ src/mutations/use-taxonomy-create-or-update-mutation.ts
@@ -1,7 +1,7 @@
 import { gql } from 'graphql-request'
 import { useRouter } from 'next/router'
 
-import { showToastNotice } from '../show-toast-notice'
+import { showToastNotice } from '../helper/show-toast-notice'
 import { useMutationFetch } from './use-mutation-fetch'
 import { TaxonomyCreateOrUpdateMutationData } from './use-set-entity-mutation/types'
 import { getRequiredString } from './use-set-entity-mutation/use-set-entity-mutation'
